enable ingredients amount for floating numbers and remove some required constraints in the create-recipe.component

+++ src/main/java/ch/pama/cookncode/domain/jpaconverter/AmountConverter.java
@@ -5,18 +5,19 @@ import ch.pama.cookncode.util.OnlyForFramework;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
+import java.math.BigDecimal;
 
 @OnlyForFramework
 @Converter(autoApply = true)
-public class AmountConverter implements AttributeConverter<Amount, Integer> {
+public class AmountConverter implements AttributeConverter<Amount, BigDecimal> {
 
     @Override
-    public Integer convertToDatabaseColumn(Amount amount) {
-        return (amount == null ? null : amount.getValue());
+    public BigDecimal convertToDatabaseColumn(Amount amount) {
+        return (amount ==  null ? null : amount.getValue());
     }
 
     @Override
-    public Amount convertToEntityAttribute(Integer amount) {
+    public Amount convertToEntityAttribute(BigDecimal amount) {
         return amount == null ? null : new Amount(amount);
     }
 }
